fix(ir): finish exact count authority enforcement
Complete the gp3.20 exact-count migration after the CountValue parser/AST foundation landed. Bump the FrankenScript language version to 2 so canonical identity changes for large and fractional count literals cannot be mistaken for v1 semantics.

Carry memory and concurrent-core authority as u64 values through IR admission, session capability tokens, the governor, and estimation. Remove the remaining f64 projection seam that could alias adjacent grants above 2^53 or reject u64::MAX after rounding to 2^64.

Keep decimal and exponent spellings exact through their canonical representation, accept only byte quantities with an exact integral scaling result, and align session resource extraction with the same checked semantics used by admission.

Extend IR and session contracts for exact-versus-fractional identity, binary-unit scaling, overflow refusal, and integer authority. Add acceptance coverage for 2^53 neighbors, u64::MAX, 2^64 refusal, exact decimal scaling, fractional-byte rejection, exponent/JSON round trips, exact core grants, and byte-for-byte token bridging.

Validation: UBS reported only test-harness panic and token-name heuristics; rustfmt, clippy, cargo check, and cargo test --no-run were clean for the eleven-file completion snapshot.

+++ crates/fs-session/src/token.rs
@@ -23,23 +23,21 @@ pub struct CapabilityToken {
     /// Wall-clock grant in seconds.
     pub wall_s: f64,
     /// Cores the session may occupy at once.
-    pub cores: f64,
+    pub cores: u64,
     /// Ledger scope (branch/namespace this session may write).
     pub ledger_scope: String,
 }
 
 impl CapabilityToken {
-    /// The static-admission view of this token. fs-ir's current capability
-    /// vocabulary represents declared memory asks as `f64`, so this is a coarse
-    /// planning projection above 2^53 bytes. The enforcing governor retains and
-    /// compares the original integer byte grant exactly.
+    /// The static-admission view of this token. Memory remains an exact `u64`
+    /// through this bridge, so static admission and the enforcing governor make
+    /// the same byte-for-byte authority decision.
     #[must_use]
     pub fn to_admission(&self) -> SessionCapability {
         SessionCapability {
             ops: self.ops.clone(),
             cores: self.cores,
-            #[allow(clippy::cast_precision_loss)] // fs-ir's admission vocabulary is f64
-            mem_bytes: self.mem_bytes as f64,
+            mem_bytes: self.mem_bytes,
             wall_s: self.wall_s,
         }
     }
